# openapi.yaml
openapi: 3.0.3
info:
  title: mjr.wtf URL Shortener API
  description: |
    A URL shortener service built with Go following hexagonal architecture principles.
    
    ## Authentication
    Most API endpoints require Bearer token authentication. Include your token in the Authorization header:
    ```
    Authorization: Bearer YOUR_TOKEN_HERE
    ```
    
    ## Rate Limiting
    The service implements rate limiting on the redirect endpoint and authenticated API routes to prevent abuse.
  version: 1.0.0
  contact:
    name: mjr.wtf API Support
  license:
    name: MIT
servers:
  - url: https://mjr.wtf
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /{shortCode}:
    get:
      summary: Redirect to original URL
      description: |
        Redirects to the original URL associated with the short code.
        This endpoint is public and does not require authentication.
        Click tracking is performed asynchronously.
      operationId: redirect
      tags:
        - urls
      parameters:
        - name: shortCode
          in: path
          description: Short code to redirect
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{3,20}$'
            example: "abc123"
      responses:
        '302':
          description: Redirect to original URL
          headers:
            Location:
              description: The original URL to redirect to
              schema:
                type: string
                format: uri
                example: "https://example.com"
        '404':
          description: Short code not found (interstitial HTML page)
          content:
            text/html:
              schema:
                type: string
                example: "404 Not Found"
        '410':
          description: Destination is marked gone (interstitial HTML page)
          content:
            text/html:
              schema:
                type: string
                example: "410 Gone"
        '429':
          description: Too many requests - rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
                minimum: 1
          content:
            text/html:
              schema:
                type: string
                example: "Too Many Requests"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Configure via AUTH_TOKENS (preferred, comma-separated) or AUTH_TOKEN (legacy single token).
        Include any active token in the Authorization header: `Authorization: Bearer YOUR_TOKEN_HERE`
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "invalid request body"
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Generic server error
              value:
                error: "internal server error"